+++ crates/web-sys/src/features/gen_VideoFrameCopyToOptions.rs
@@ -15,6 +15,42 @@ extern "C" {
     #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
     pub type VideoFrameCopyToOptions;
     #[cfg(web_sys_unstable_apis)]
+    #[doc = "Get the `colorSpace` field of this object."]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `VideoFrameCopyToOptions`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
+    #[wasm_bindgen(method, getter = "colorSpace")]
+    pub fn get_color_space(this: &VideoFrameCopyToOptions) -> Option<::alloc::string::String>;
+    #[cfg(web_sys_unstable_apis)]
+    #[doc = "Change the `colorSpace` field of this object."]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `VideoFrameCopyToOptions`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
+    #[wasm_bindgen(method, setter = "colorSpace")]
+    pub fn set_color_space(this: &VideoFrameCopyToOptions, val: &str);
+    #[cfg(web_sys_unstable_apis)]
+    #[doc = "Get the `format` field of this object."]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `VideoFrameCopyToOptions`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
+    #[wasm_bindgen(method, getter = "format")]
+    pub fn get_format(this: &VideoFrameCopyToOptions) -> Option<::alloc::string::String>;
+    #[cfg(web_sys_unstable_apis)]
+    #[doc = "Change the `format` field of this object."]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `VideoFrameCopyToOptions`*"]
+    #[doc = ""]
+    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
+    #[doc = "[described in the `wasm-bindgen` guide](https://wasm-bindgen.github.io/wasm-bindgen/web-sys/unstable-apis.html)*"]
+    #[wasm_bindgen(method, setter = "format")]
+    pub fn set_format(this: &VideoFrameCopyToOptions, val: &str);
+    #[cfg(web_sys_unstable_apis)]
     #[doc = "Get the `layout` field of this object."]
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `VideoFrameCopyToOptions`*"]
@@ -65,6 +101,18 @@ impl VideoFrameCopyToOptions {
         #[allow(unused_mut)]
         let mut ret: Self = ::wasm_bindgen::JsCast::unchecked_into(::js_sys::Object::new());
         ret
+    }
+    #[cfg(web_sys_unstable_apis)]
+    #[deprecated = "Use `set_color_space()` instead."]
+    pub fn color_space(&mut self, val: &str) -> &mut Self {
+        self.set_color_space(val);
+        self
+    }
+    #[cfg(web_sys_unstable_apis)]
+    #[deprecated = "Use `set_format()` instead."]
+    pub fn format(&mut self, val: &str) -> &mut Self {
+        self.set_format(val);
+        self
     }
     #[cfg(web_sys_unstable_apis)]
     #[deprecated = "Use `set_layout()` instead."]